retrace,glretrace: Call profiling option. Hook up the Metrics abstraction to profiling calls.
New option is "glretrace --pcalls".
It profiles all calls (that have side effects) based on the
backends in glretrace::metricBackends and metrics selected in them (via
enableMetric() backend interface call). CLI interface into this
selection is to be done in future commits.
Backends are used in series. For each backend required number of passes
is evaluated.

It does not work with old profiling options.