retrace,glretrace: Option for metrics listing.
The option is "--list-metrics".
It outputs metrics and metric groups for each plugged backend.

It also requires the trace file to be specified, as some backends
can query metrics only after context is created. They can also possibly
depend on the context version.
Trace file is only replayed until the context is created.