retrace,glretrace: Frame profiling option. Hook up the Metrics abstraction to profiling frames.
New option is "glretrace --pframes".
It profiles all frames based on the
backends in glretrace::metricBackends and metrics selected in them (via
enableMetric() backend interface call). CLI interface into this
selection is to be done in future commits.
Backends are used in series. For each backend required number of passes
is evaluated.

It does not work with old profiling options.