"common" backend for the Metrics abstraction.
Backend implementation in: metric_backend_common.hpp
                           metric_backend_common.cpp
This backend offers the same functionality as currently available flags
for profiling (--pcpu, --pgpu etc.)
It reuses the old code for actual profiling most of the time. The only difference is that it uses GL_TIMESTAMP for determining gpu
call duration if available (so that queries can be nested).
Some metrics are profiled in two passes if requested for frames as well
as calls.