Add custom allocator that uses mmap. Also make metric_writer use it.
Profiling uses a lot of memory as all data is kept through the whole
run. This is an attempt to reduce the usage.
MmapAllocator creates temporary files on disk and mmaps them.
It can be used by STL containers.
Currently enabled only on unix-like systems.
The idea is to use this allocator in writer and backends where otherwise
data significantly increases RAM usage.